se hacen areeglos y mejoras en el codigo para un mejor funcionamiento
