+++ src/auth/auth.module.ts
@@ -1,25 +1,26 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-import { RolesGuard } from './roles.guard';
-import { JwtStrategy } from './jwt.strategy';
+import { AuthService } from './services/auth.service';
+import { AuthController } from './controllers/auth.controller';
+import { RolesGuard } from './guards/roles.guard';
+import { JwtStrategy } from './strategies/jwt.strategy';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { AdminModule } from 'src/modules/admin/admin.module';
 import { UsersModule } from 'src/modules/users/users.module';
+import { UsersService } from 'src/modules/users/services/users.service';
 
 @Module({
   imports: [
     PassportModule,
     JwtModule.register({
-      secret: process.env.JWT_SECRET, // Reemplaza con tu propia clave secreta
-      signOptions: { expiresIn: '60m' }, // Configura la duración del token JWT
+      secret: process.env.JWT_SECRET,
+      signOptions: { expiresIn: process.env.ACCESS_TOKEN_EXPIRY || '15m' },
     }),
     AdminModule,
     UsersModule,
   ],
   controllers: [AuthController],
   providers: [AuthService, JwtStrategy, RolesGuard],
-  exports: [AuthService],
+  exports: [AuthService, JwtStrategy, JwtModule, PassportModule],
 })
 export class AuthModule {}
