+++ src/auth/guards/roles.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
-import { Role } from './auth.type';
+import { Role } from '../types/auth.type';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
