+++ src/auth/strategies/jwt.strategy.ts
@@ -3,16 +3,21 @@ import { PassportStrategy } from '@nestjs/passport';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
+export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
   constructor() {
     super({
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
       ignoreExpiration: false,
-      secretOrKey: 'your-secret-key', // Reemplaza con tu propia clave secreta
+      secretOrKey: process.env.JWT_SECRET,
     });
   }
 
   async validate(payload: any) {
-    return { username: payload.username, role: payload.role };
+    return {
+      username: payload.username,
+      role: payload.role,
+      email: payload.email,
+      password: payload.password,
+    };
   }
 }
