+++ src/main.ts
@@ -1,6 +1,10 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+// import * as dotenv from 'dotenv';
+
+// Cargar variables de entorno
+// dotenv.config();
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
