--- src/modules/admin/services/admin.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AdminService } from './admin.service';
-
-describe('AdminService', () => {
-  let service: AdminService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AdminService],
-    }).compile();
-
-    service = module.get<AdminService>(AdminService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
